app/requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
fpdf
plotly
google-generativeai
